+++ lib/plausible/stats/legacy/legacy_query_builder.ex
@@ -93,7 +93,7 @@ defmodule Plausible.Stats.Legacy.QueryBuilder do
   end
 
   defp put_period(query, site, %{"period" => period} = params)
-       when period in ["7d", "28d", "30d", "90d"] do
+       when period in ["7d", "28d", "30d", "91d"] do
     {days, "d"} = Integer.parse(period)
 
     end_date = parse_single_date(query, params) |> Date.shift(day: -1)