+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Ability to sort by and compare the `exit_rate` metric in the dashboard Exit Pages > Details report
 - Add top 3 pages into the traffic spike email
-- Two new shorthand time periods `28d` and `90d` available on both dashboard and in public API
+- Two new shorthand time periods `28d` and `91d` available on both dashboard and in public API
 - Average scroll depth metric
 - Scroll Depth goals
 - Dashboard shows comparisons for all reports